# Skip "already done" tasks -- 10x faster restart