--- a/src/pypeflow/controller.py
+++ b/src/pypeflow/controller.py
@@ -488,7 +488,6 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
         self.thread_handler = thread_handler
         self.messageQueue = messageQueue
         self.shutdown_event = shutdown_event
-        self.jobStatusMap = {}
 
     def addTasks(self, taskObjs):
         """
@@ -554,7 +553,7 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
 
         sortedTaskList = [ (str(u), self._pypeObjects[u], self._pypeObjects[u].getStatus()) for u in tSortedURLs
                             if isinstance(self._pypeObjects[u], PypeTaskBase) ]
-        self.jobStatusMap = dict( ( (t[0], t[2]) for t in sortedTaskList ) )
+        jobStatusMap = dict( ( (t[0], t[2]) for t in sortedTaskList ) )
 
         prereqJobURLMap = {}
 
@@ -576,6 +575,7 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
         loopN = 0
         lastUpdate = None
         activeDataObjs = set() #keep a set of output and mutable data object. a task can not be submitted if a running task has the same output
+        updatedTaskURLs = set() #to avoid extra stat-calls
 
         while 1:
 
@@ -589,7 +589,7 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
             jobsReadyToBeSubmitted = []
 
             for URL, taskObj, tStatus in sortedTaskList:
-                if self.jobStatusMap[URL] != TaskInitialized:
+                if jobStatusMap[URL] != TaskInitialized:
                     # TODO(CD): Can mutableDataObjs.values() change between iterations?
                     #logger.debug('Found %s updated in jsmap' %URL)
                     continue
@@ -597,8 +597,8 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
 
                 logger.debug(' preqs of %s:' %URL)
                 for u in prereqJobURLs:
-                    logger.debug('  %s: %s' %(self.jobStatusMap[u], u))
-                if any(self.jobStatusMap[u] != "done" for u in prereqJobURLs):
+                    logger.debug('  %s: %s' %(jobStatusMap[u], u))
+                if any(jobStatusMap[u] != "done" for u in prereqJobURLs):
                     # Note: If jobStatusMap[u] raises, then the sorting was wrong.
                     #logger.debug('Prereqs not done! %s' %URL)
                     continue
@@ -623,16 +623,17 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
                     if outputCollision:
                         continue
 
+                # We use 'updatedTaskURLs' to short-circuit 'isSatisfied()', to avoid many stat-calls.
                 # Note: Sorting should prevent FileNotExistError in isSatisfied().
-                if taskObj.isSatisfied():
+                if not (set(prereqJobURLs) & updatedTaskURLs) and taskObj.isSatisfied():
                     #taskObj.setStatus(pypeflow.task.TaskDone) # Safe b/c thread is not running yet, and all prereqs are done.
-                    logger.debug(' Skipping already done task: %s (Status was %s)' %(URL, self.jobStatusMap[str(URL)]))
+                    logger.debug(' Skipping already done task: %s (Status was %s)' %(URL, jobStatusMap[str(URL)]))
                     taskObj.setStatus(TaskDone) # to avoid re-stat on subsequent call to refreshTargets()
-                    self.jobStatusMap[str(URL)] = TaskDone # to avoid re-stat on *this* call
+                    jobStatusMap[str(URL)] = TaskDone # to avoid re-stat on *this* call
                     successfullTask = self._pypeObjects[str(URL)]
                     successfullTask.finalize()
                     continue
-                self.jobStatusMap[str(URL)] = "ready" # in case not all ready jobs are given threads immediately, to avoid re-stat
+                jobStatusMap[str(URL)] = "ready" # in case not all ready jobs are given threads immediately, to avoid re-stat
                 jobsReadyToBeSubmitted.append( (URL, taskObj) )
                 for dataObj in taskObj.outputDataObjs.values() + taskObj.mutableDataObjs.values():
                     logger.debug( "add active data obj:"+str(dataObj))
@@ -645,8 +646,8 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
             if numAliveThreads == 0 and len(jobsReadyToBeSubmitted) == 0 and self.messageQueue.empty(): 
                 logger.info( "_refreshTargets() finished with no thread running and no new job to submit" )
                 for URL in task2thread:
-                    assert self.jobStatusMap[str(URL)] in ("done", "fail"), "status(%s)==%r" %(
-                            URL, self.jobStatusMap[str(URL)])
+                    assert jobStatusMap[str(URL)] in ("done", "fail"), "status(%s)==%r" %(
+                            URL, jobStatusMap[str(URL)])
                 break # End of loop!
 
             for URL, taskObj in jobsReadyToBeSubmitted:
@@ -659,7 +660,7 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
                     nSubmittedJob += 1
                     usedTaskSlots += taskObj.nSlots
                     numAliveThreads += 1
-                    self.jobStatusMap[URL] = "submitted"
+                    jobStatusMap[URL] = "submitted"
                     # Note that we re-submit completed tasks whenever refreshTargets() is called.
                     logger.debug("Submitted %r" %URL)
                     logger.debug(" Details: %r" %taskObj)
@@ -681,7 +682,8 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
             while not self.messageQueue.empty():
                 sleep_time = 0 # Wait very briefly while messages are coming in.
                 URL, message = self.messageQueue.get()
-                self.jobStatusMap[str(URL)] = message
+                updatedTaskURLs.add(URL)
+                jobStatusMap[str(URL)] = message
                 logger.debug("message for %s: %r" %(URL, message))
 
                 if message in ["done"]:
@@ -710,18 +712,17 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
                 elif message in ["started, runflag: 0"]:
                     logger.debug("Queued %s (already completed) ..." %repr(URL))
                     raise Exception('It should not be possible to start an already completed task.')
-
                 else:
                     logger.warning("Got unexpected message %r from URL %r." %(message, URL))
 
-            for u,s in sorted(self.jobStatusMap.items()):
+            for u,s in sorted(jobStatusMap.items()):
                 logger.debug("task status: %r, %r, used slots: %d" % (str(u),str(s), self._pypeObjects[str(u)].nSlots))
 
             if failedJobCount != 0 and exitOnFailure:
                 raise TaskFailureError("Counted %d failures." %failedJobCount)
 
 
-        for u,s in sorted(self.jobStatusMap.items()):
+        for u,s in sorted(jobStatusMap.items()):
             logger.debug("task status: %s, %r" % (u, s))
 
         self._runCallback(callback)
@@ -753,7 +754,7 @@ class _PypeConcurrentWorkflow(PypeWorkflow):
                     s = URLParseResult.scheme + "://..." + URLParseResult.path.split("/")[-1] 
 
                 if URLParseResult.scheme == "task":
-                    jobStatus = self.jobStatusMap.get(URL, None)
+                    jobStatus = jobStatusMap.get(URL, None)
                     if jobStatus != None:
                         if jobStatus == "fail":
                             color = 'red'