@pb-cdunn 2015-11-23
With the latest change, stat-calls are avoided when any prereq was updated, since we *know* we want to re-run. That reduces stat-calls from 3008 to 2639, vs. 2242 before any changes. I don't know what was skipped before, but I suspect there was some kind of bug.

Now, on any call to refreshTargets:
- A task is skipped if its state is "done".
- Otherwise, when all prereqs are "done" ...
  - If any prereq was updated *on that call*, then the task is run.
    - (This isn't just a matter of a few stat-calls. Eventually, when we have a daligner caching-wrapper, we will want daligner to re-run always. But other steps will still rely on timestamps. So we'll force re-running when a previous step was re-run.)
  - But if not, then files are compared before deciding to run.

Also, all stat-calls are now on the main thread. That prevents a thread from being created unless a task actually needs to be run.

Old (in lustre):
```
clean run:
real    1m9.937s
user    1m10.221s
sys     5m46.438s

restart:
real    0m28.046s
user    0m19.124s
sys     0m11.589s
```
New (in lustre):
```
clean run:
real    0m53.935s
user    0m40.265s
sys     1m6.316s
re-start:
real    0m3.856s
user    0m2.740s
sys     0m0.611s
```
(I think the 'sys' times are heavily influenced by the freshness of the file-system cache.)

Fixing doctests now...