@pb-cdunn 2015-11-23
Doctests pass. Everything works for me.