@pb-cdunn 2015-11-23
All tests pass. I'll merge this later tonight.