@pb-jchin 2015-11-24
should we bump the minor version? 

I am also interested in testing out a database backed DataObject solution. I may do some prototyping during the holiday season.