@pb-cdunn 2015-11-24
We're using an integration-package to select versions, so we don't need library-level versioning. But it doesn't really matter to me. A bump is fine if you want. A case could be made for major, minor, or minor, depending on point-of-view. I'll try to summarize any user-level changes in the merge-message.

I'll comment separately on the idea of a database.